refactor: functions moved to _base.py
