+++ tests/database_builder_libs/stores/typedb/test_query.py
@@ -1,8 +1,5 @@
 import pytest
-from database_builder_libs.stores.typedb._query import (
-    _validate_identifier,
-    _escape_string,
-)
+from database_builder_libs.stores.typedb._base import _validate_identifier, _escape_string
 
 
 class TestValidateIdentifier:
